The sending of double values delegated in the protocol

+++ src/server/server_DynamiteNotification.cpp
@@ -1,6 +1,6 @@
 #include "server_DynamiteNotification.h"
 
-DynamiteNotification::DynamiteNotification(ClientHandler* client, int x, int y,\
+DynamiteNotification::DynamiteNotification(ClientHandler* client, double x, double y,\
                                                 int time_to_explosion){
   this->client = client;
   this->x = x;
