+++ src/server/server_RadiocontrolledNotification.cpp
@@ -1,7 +1,7 @@
 #include "server_RadiocontrolledNotification.h"
 
 RadiocontrolledNotification::RadiocontrolledNotification(ClientHandler* client,\
-                                                size_t id, int x, int y){
+                                                size_t id, double x, double y){
   this->client = client;
   this->id = id;
   this->x = x;
