+++ src/server/server_TeletransportationEvent.cpp
@@ -1,6 +1,6 @@
 #include "server_TeletransportationEvent.h"
 
-TeletransportationEvent::TeletransportationEvent(Player* player, int x, int y){
+TeletransportationEvent::TeletransportationEvent(Player* player, double x, double y){
   this->player = player;
   this->x = x;
   this->y = y;
