+++ src/server/server_WormNotification.h
@@ -8,8 +8,8 @@ private:
     ClientHandler* client;
     size_t id;
     size_t life_points;
-    int x;
-    int y;
+    double x;
+    double y;
     int angle;
     bool is_facing_right;
     int team;
@@ -18,7 +18,7 @@ private:
     WormNotification& operator=(const WormNotification&) = delete;
 public:
     WormNotification(ClientHandler* client, size_t id, size_t life_points,\
-                      int x, int y, int angle, bool is_facing_right, int team);
+                      double x, double y, int angle, bool is_facing_right, int team);
 
     void run();
